chore: Update import paths for components in code changes
