+++ README.md
@@ -1,0 +1,16 @@
+
+### Resources 
+
+#### Pipeline Resources
+
+- [EAS CLI](https://github.com/expo/eas-cli?tab=readme-ov-file#eas-channeledit-name)
+
+- [Expo Github Action](https://github.com/expo/expo-github-action?tab=readme-ov-file#create-previews-on-prs)
+  
+- [Slack Notification](https://github.com/marketplace/actions/slack-notification)
+
+- [QR Code API](https://goqr.me/api/doc/create-qr-code/)
+  
+- [gh-get-current-pr](https://github.com/marketplace/actions/get-current-pull-request)
+
+- [sticky-pull-request-comment](https://github.com/marocchino/sticky-pull-request-comment)
