+++ src/components/PaymentBar.tsx
