+++ src/components/onboarding/Progressbar.tsx
