+++ src/screens/LoginScreen.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { View, TextInput, Button } from "react-native";
 import { Auth } from "aws-amplify";
-import LoginComponent from "../../components/login/LoginComponent";
+import LoginComponent from "../components/login/LoginComponent";
 
 const LoginScreen = ({ navigation }) => {
   const [email, setEmail] = useState("");
