+++ src/screens/NotFoundScreen.js
@@ -1,7 +1,7 @@
 import { Link, Stack } from "expo-router";
 import { StyleSheet } from "react-native";
 
-import { Text, View } from "@/components/Themed";
+import { Text, View } from "../components/Themed";
 
 export default function NotFoundScreen() {
   return (
