+++ src/utils/helperFunctions.ts
